Unified binary generating parameters for Poseidon 1 and 2.
This is base work for poseidon2-gl.
