+++ plonky3/Cargo.toml
@@ -10,12 +10,8 @@ repository.workspace = true
 name = "powdr_plonky3"
 
 [[bin]]
-name = "gen_poseidon_bb_consts"
-path = "src/bin/gen_poseidon_bb_consts.rs"
-
-[[bin]]
-name = "gen_poseidon2_bb_consts"
-path = "src/bin/gen_poseidon2_bb_consts.rs"
+name = "gen_poseidon_consts"
+path = "src/bin/gen_poseidon_consts.rs"
 
 [features]
 default = []
